requirements.txt
litellm
python-dotenv
faiss-cpu
sentence-transformers
numpy
